# internal/transport/http/handlers/movie/get_movie.go
package movie

import (
	"encoding/json"
	"net/http"
	"strconv"
)

// getMovieHandler godoc
// @Summary Get movie
// @Description get movie
// @Tags movies
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param Authorization header string true "Authorization"
// @Param id path int true "movie ID"
// @Success 200 {object} getMovieResponse "Success"
// @Failure 400 {object} errorResponse "Bad request"
// @Failure 401 {object} errorResponse "Unauthorized"
// @Failure 500 {object} errorResponse "Internal server error"
// @Router /movies/{id} [get]
func (c *Controller) getMovieByIDHandler(w http.ResponseWriter, r *http.Request) {
	idStr := r.PathValue("id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		sendErrorResponse(w, err.Error(), http.StatusBadRequest)
		return
	}

	movie, err := c.moviesService.GetMovie(r.Context(), id)
	if err != nil {
		sendErrorResponse(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	err = json.NewEncoder(w).Encode(convertMovie(movie))
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
}